requirements.txt
gradio>=4.0.0,<7.0.0
langchain-openai
langchain-community
langgraph
wikipedia
python-dotenv
